+++ CHANGES.md
@@ -1,5 +1,7 @@
 #### [unreleased]
 * refactored for traits, now requires PHP 5.4
+* updated for new `Singleton` factory
+* updated for GitLab API v4
 
 #### 1.1.0 / 2017-10-30
 * refactored `class API` to not be dependent on `class Base`